Add the rest of the pages and layouts for export
