+++ src/lib/pages/workflow-history-feed.svelte
@@ -1,0 +1,6 @@
+<script lang="ts">
+  import EventSummary from '$lib/components/event/event-summary.svelte';
+  import { events, eventGroups } from '$lib/stores/events';
+</script>
+
+<EventSummary items={$events} groups={$eventGroups} />
