+++ src/lib/pages/workflow-history-json.svelte
@@ -1,0 +1,6 @@
+<script lang="ts">
+  import CodeBlock from '$lib/components/code-block.svelte';
+  import { events } from '$lib/stores/events';
+</script>
+
+<CodeBlock content={$events} data-cy="event-history-json" />
