+++ src/routes/namespaces/[namespace]/workflows/[workflow]/[run]/history/compact/index.svelte
@@ -1,10 +1,5 @@
 <script lang="ts">
-  import EventSummary from '$lib/components/event/event-summary.svelte';
-  import { ascendingEventGroups } from '$lib/stores/events';
+  import WorkflowHistoryCompact from '$lib/pages/workflow-history-compact.svelte';
 </script>
 
-<EventSummary
-  items={$ascendingEventGroups}
-  groups={$ascendingEventGroups}
-  compact={true}
-/>
+<WorkflowHistoryCompact />
